[block] Fix race conditions in Block I/O protocol
- BUG: In blockserver_fifo_complete, the following events
occur:
  1) LOCK
  2) Complete 'msg->txn', making it usable again
  3) UNLOCK
  4) Nullify 'msg->txn' deleting it.
Unfortunately, if between (3) and (4), the 'msg->txn' entry (originally
allocated from 'msgs_', within BlockTransaction::Enqueue) is
re-allocated, then when (4) occurs, it nullifies a transaction which
expects to access the now-null 'msg->txn' during a callback.
- FIX: Swap steps (3) and (4).
- Add clang thread safety analyzers.
- Added some debug asserts, to catch similar issues in the future (if a
refactor occurs). Running blktest in a loop on the SATA driver was
sufficient to find this race.

Change-Id: I05ee76a998c9ec25ad4574e863a8b6e696d007c7