[hypervisor] Add bi-directional guest control.
Change the serial FIFO to be a control FIFO, so that we can control
the execution of guests through packets passed between the hypervisor
and user space.

Currently this is being used to pass serial output back to user space,
but it will soon be used to handle EPT violations within guests.

Change-Id: I862e8e062832a0a5ab176d25bea0af839c6a4692