[framebuffer] dumb framebuffer common core driver
Only built if ENABLE_NEW_FB is true, to avoid fighting
with gfxconsole until gfxconsole is ready to use this.

This driver binds against any Protocol Display devices
and publishes a device under /dev/class/framebuffer which
speaks a combination of Display and Console ioctls similar
to the gfxconsole's vc nodes.

Each open of this device returns an instance which can provide
an offscreen framebuffer and the ability to flush that offsceen
framebuffer to the underlying hw framebuffer.

gfxfractal is updated to accept a device name so this can be
tested with: gfxconsole /dev/class/framebuffer/000

gfxconsole is updated to build as a misc device when ENABLE_NEW_FB
is true, and currently does nothing in that mode.

MG-774

Change-Id: I888ba1501d9f0d68babc6eabac21f4a8ed518d09