Refactor the use of the throttle output.
Turn the already fairly special cased throttle output into a full-on
special case.  IOW - in the OutputManager, do not add the throttle
output to the set of active outputs, and hold a dedicated reference to
it.  In renderer's, do not hold the link to the throttle output in the
set of active output links, hold a dedicated reference to the link to
the throttle output.

Refactoring this way makes it a bit easier to implement the temporary
simplified audio routing policies because the throttle output is not
considered to be part of the set of active outputs when making
decisions about which renderers should be assigned to which outputs.

Also, when linking a renderer to a new output, having a dedicated
reference to the throttle output link means that we don't have to go
and find the throttle output link when we want to initialise the
pending output queue of the new link.

Change-Id: I3cfa0fecba0b5deecdf13d25f1b94e89330f4dbb