Use hardlink for copy command
This has a different semantics, but the way "copy" tool is being used
is compatible with links (where later changes to the source file
implicitly change the destination file), since Ninja is tracking
changes to the source file and any change would trigger a rebuild.
This is much faster and more efficient, especially for large files.

Change-Id: I0db1c0a201dc13978a407d8009823510c917c3f2