[kernel][thread exit] move final kernel thread teardown to a DPC
Previously, the kernel would complete the freeing of resources of a
thread by putting the stack(s) and thread_t structure on a 'delayed free
list' which simply delays freeing until the next caller of any of the
heap functions.

This mechanism relied on an implicit serialization with the heap's mutex
when it grabbed the thread_lock. Since the mutex now no longer grabs the
thread_lock in the uncontended case, it was possible for another thread
to race the delayed free and free the stack/thread_t before the original
thread had completed a reschedule.

Move the delayed free to a dpc, which then has it's own serialization
hack to make sure it doesn't race with the exiting thread.

Needed to add a few more specialized functions to dpc and event to allow
queuing a dpc with the thread_lock held.

MG-781 #done
MG-779 #done
MG-778 #done

Change-Id: I21fe0ccc8b5da60dd4fa04ad7245505b437b9da5