[syscalls][socket] use mbuf-style backing memory
This switches the kernel data structure used to store pending
reads and writes for sockets from a single large VMO to a
chain of small buffers, allocated as needed.

The inspiration for this is the traditional mbuf linked-list seen
in other kernel socket implementations.

As this mbuf is not used by an in-kernel network stack, it
comes out a lot smaller than other kernel mbufs. It is a
singly-linked list instead of doubly-linked, and it has no
notion of chains. It will probably acquire a tiny header
space for UDP datagram support later, along with a datagram
size field.

Change-Id: Iaedce8f970b7e72eb22ad408761d9e5cbe79ab71