apps/media: change formatting endl policy in flog
The previous policy used in operator<< overloads for formatting in
flog was to terminate multiline output with std::endl and not to
do that for single-line output. This makes a little bit of sense,
but creates a situtation where we have to know which behavior we're
getting every time we format something.

The new policy is to never terminate the last line. This makes all
the operators consistent. Of course, we'd never want to put anything
other that std::endl after a multiline item, but oh well.

Change-Id: I3bafa4da74afd371156c0b75e9873b38aa1b8e9c