[kernel][arm64] Add adr_global macro, use it
The new macro encapsulates the standard ADRP, ADD sequence for
materializing the address of a global symbol into a register.

Also change uses of a literal pool to materialize a normal address to
use adr_global.  The remaining uses of load from literal pool should
be confined to the cases that need to load an unadjusted compile-time
constant rather than computing from the runtime PC.

Change-Id: I368b5544e008bfd0e10d3900184ccb7091031e98