[mdi] Use virt addresses in mdi
We were using physical addresses in mdi for peripheral base
addresses, then converting to virt via paddr_to_kvaddr.  This
Caused the base addresses to be in the 0xffff000000000000 mapping
instead of the typical 0xffffffffc0000000 peripheral mapping and
therefore the accesses would be cached.  This had not been a problem
previously as the only target which would encounter the problem
was qemu-arm64, which does not emulate cache behavior.

This change switches to using virtual peripheral addresses instead
of base addresses.  This should be cleaned up down the road when
we move the address mappings to mdi since with this change it will
require the mdi files to be kept consistent with the
target/xxxxx/rules.mk definitions PERIPH_BASE_VIRT.

Change-Id: I326bad803a2554e9e9f93d66bd104610bd9e8235