[merkle] Change size_t to uint64_t
This CL modifies fields with the type of size_t to have the type of
uint64_t instead.  The exact number of bits in these fields have direct
impacts on the Merkle tree (e.g. the max alloawed height of a tree, the
construction of the 'locality' prefix, etc.).  As a result, it is
clearer and more correct to be explicit in the number of bits used.

Change-Id: I9da0d04bdca0cf60beb2e74b1cc2554bdb085c87