[merkle] Refactor tree verification.
This CL reworks Verify, VerifyRoot, and VerifyLevel to be clearer with
less "magic state" being carried around.  In particular, offsets_ and
ranges_ are no longer used, and the digest behavior is captured in
specific helper methods.  These changes allow VerifyRoot and VerifyLevel
to also become static methods. Finally, a few off-by-one errors that
were revelaed by the refactor are fixed.

Change-Id: I071326ec76a00a4764de13a804fa3961586a17ce