[merkle] Refactor C wrappers
This CL updates the C wrapper function for merkle::Tree.  It normalizes
the function names to match more closely, i.e. merkle::Tree::SomeMethod
has a corresponding C wrapper named merkle_tree_some_method.

It also no longer manages the Merkle tree memory transparently, but
requires it to be passed in as with the C++ API.  This is to allow the
tree to be something more complicated than just a pre-allocated buffer,
e.g. a VMO with on demand paging.

Finally, it rolls back some of the earlier size_t->uint64_t conversion
for method signatures.  The notions of "levels" and "localities" are
kept as explicitly 64 bits due to the role in the hash; all others have
been made size_t's.

Change-Id: Ia4a72ddaeb1d70232589c228a2dfbe26ef749b49