Prevent overriding an already existing commit in the cloud.
If 2 devices produce the same merge commit, they will both try to upload
it. Because of firebase implementation, the second one will erase the
timestamp from the first one. If the first one had created another
commit as a child of the merge commit, this commit might have a
timestamp on the cloud that is smaller than its parent, which break the
Ledger invariants.

To solve this we do the following:
- On firebase, we prevent client to override an existing commit.
- On the client, before uploading a commit, we erase any already synced
  commit.

It means that if 2 devices try to upload the same merge commit, the
second one won't be able to upload to the cloud. Instead, when it
retries, the commit will be removed from its queue because it will have
received it from the cloud.

LE-214 #done Fixed

Change-Id: I98306ad71720c7ea6caf49c9704c4b6accffe8d8