[udev][usb-xhci] Fix hang when attempting to transfer maximum transfer size
In xhci_get_max_transfer_size(), we compute the maximum amount of data
we can read or write based on the number of TRBs in our transfer ring assuming
worst case scenario where the buffer is completely fragmented and we can only
transfer one page per TRB. However we forgot to take into account the LINK TRB
at the end of transfer ring, so we had an off by one error.
In the case where we had a completely fragmented buffer that was of max transfer size
in length, we were one TRB short and was not able to queue an event data TRB
which is needed to notify us when the transfer is complete.
Now we subtract one to take the LINK TRB into account so we do not run into this situation.

Fixes at least part of US-249

Change-Id: I0231c5982c8d45f79a6a6de37b12fc765df59bd2