Harden the reset sequence and DMA stream managment.
Be much more conservative about how we initialize the driver in case
we are coming back from a driver crash.  In particular...

++ Do not attempt to reset the controller while there are stream
   descriptor DMA engines being held in reset.
++ Do not leave stream descriptors being held in reset at all times.
   Just suffer the time it takes to set them up completely every time
   we need to use one.

Failure to do these things seems to lead to some Bad Things, starting
with not recovering properly from a reset and eventually leading to
wedging the controller hardware so bad that it no longer responds to
the HW reset cycle and the device needs to be powered down, have power
removed for 15 seconds, and the cold booted (rebooting does not
unstick the hardware).

MG-782 #done

Change-Id: I0bc7a687821c0e382f8647ae06772421747f061e