apps/media: use TimelineControlPoint in video renderer
The video renderer (actually VideoFrameSource) was implementing the
MediaTimelineControlPoint and MediaTimelineConsumer interfaces. We
have a utility class (TimelineControlPoint) that also does that. This
CL changes VideoFrameSource to use the utility class.

Besides getting rid of some duplicate code, this change allows us to
instrument the timeline control point separately. Currently, the
activity of the timeline control point is recorded using the
MediaRenderer's channel, which is a problem for the audio renderer,
which already uses TimelineControlPoint.

There will be subsequent CL that changes the instrumentation. This
CL breaks instrumentation in that the timeline control point activity
is no longer logged.

Change-Id: I50e64341a80809ffefeba1fbe0cdaf0e2a9171ce