[command_line] Use StringView for querying options
This converts option lookups to use StringView instead of const
std::string& to allow for more efficient lookups in some situations,
such as querying from pieces of a larger string. The string values
themselves are stored in CommandLine::Option member variables inside
std::string instances and functions that return ownership of the string
still use std::string.

We also have to do a copy from the StringView to a std::string instance
in order to do a map lookup which is an unfortunate consequence of C++
unordered containers not having any support for heterogenuous lookups.

Change-Id: I9a6ce5b30496273f20126d6d5d5b57bc83c460b7