[udev][framebuffer] ownership, events. groups
The framebuffer device instances now belong to one of two groups,
"virtcon" or "fullscreen".  Later we may introduce more groups.

Only one framebuffer client (a fullscreen app) may be active
at a time.

The virtcon may maintain multiple clients (which it uses for
multiple virtual framebuffers), and may select which group (virtcon
or framebuffer) is active and able to repaint the underlying
framebuffer.

Next: detect gpu-enabled displays and treat the swap chain as group 1
for those.

Change-Id: Ic0a244a6bf333034912993bb9ab9a6fc69450064