[udev][i8042] Fix failures in device_add()
As a side effect of our recent change that eliminates mx_driver_t
from the DDK APIs, we must now create a device in the driver bind
op when binding to the MISC_PARENT root. Otherwise, device_add()
is unable to figure out which driver the device belongs to.
Since the i8042 driver creates its devices later in a thread,
we must now create a root device in i8042_bind() for the driver
to continue to work

Change-Id: I85bca44a6ea88f3b3e5e18aa04613b301926f03b