Fix rapidjson crash and delete creds db file on parsing failure.
Simplified creds.db file structure that gets rid of nested json objects
that is causing the crash. The new creds.db file is a json object with
account_id and the corresponding refresh_token.

Also only the in-memory cache was getting reset on data corruption. This cl
also resets the credentials file on disk in such failure cases.

Change-Id: Ia2590aba49b1feb2a2e639c5640f05a938d54bca