rpc: split server creation and servicing
Users can now catch initialization errors in a setup goroutine and then just
spawn a goroutine for the dispatcher loop. This change also makes it possible
to write tests that can reliably bind a server and then start using it
without having to probe or sleep to see if the server goroutine has finished
initialization.

Change-Id: I40954ff7a26f60af169d54fbe6fd316983b571fe