rpc: split server creation and servicing
Users can now catch initialization errors in a setup goroutine and then just
spawn a goroutine for the dispatcher loop. This change also makes it possible
to write tests that can reliably bind a server and then start using it
without having to probe or sleep to see if the server goroutine has finished
initialization.

Change-Id: Idf07688394bdd8020860e29d11f9d2ed6c181541