[kernel][x86] Prevent use of a non-canonical return address with SYSRET
Before this change, it was possible to get SYSRET to return to a
non-canonical return address by using mx_thread_write_state() on a
thread that was suspended while inside a syscall.  That is exploitable
(see sysret_problem.md).

Fix that by disallowing non-canonical addresses in this case.

Thanks to Jann Horn for pointing out this bug.

MG-810

Change-Id: Ieca21674eb5e404fbb8ebc2dada562bd8b688cdf