Generate helper methods for Rust
Creating a channel and then binding a client and server to the ends is a
very common idiom when using the Rust fidl bindings. Unfortunately the
fact that the different structs are tied together only by a name prefix
means it can't be abstracted into a macro or function.

This generates helper functions for each interface to make creating a
request and a proxy easier.

Some methods take a channel instead of a Server, so I also added a
convenience method for converting a Server to a channel. Previously
the only way to do this was Channel::from_handle(server.into_handle())
and the round trip seemed redundant given that the Server contained a
Channel.

Change-Id: I0df9cccc62e70db35739cfd2b1d8b2623a3c67d2