[gfxconsole] unravel locking a bit, support show/hide log at start
The vc_*() api surface is pretty sprawling.  Instead of trying to
having some of it hold locks and some of it assume locks are held,
move to a regime where the threads at the bottom of main.cpp acquire
a lock before calling into *any* vc_*() interfaces.

Removed the vc_ prefix from the thread functions at the bottom of
main to make it clear they're not part of the single threaded vc_*()
core code.

Eventually these threads should all collapse into a single ports-based
event processor and we'll be able to drop the lock entirely.

Change-Id: I2c57c884644d17f6aacab2c8e1a1e626fdff796d