apps/netconnector: use own ServiceProviderImpl for NetStubResponder
NetStubResponder is a lib class that registers a service provider
with NetConnector and instantiates stubs that allow remote parties
to control a local service implementation. NetStubResponder was
adding the service name with the process-global 'outgoing services'
ServiceProviderImpl and binding that to the channel it passes to
NetConnector when registering. The effect was that the provision
callback in NetStubResponder could be called after the responder
had been deleted.

This CL gives NetStubResponder its own ServiceProviderImpl with
which to do the registration. That ServiceProviderImpl is closed
in the responder's destructor, effectively unregistering from
NetConnector and preventing the provision callback from being
called.

Change-Id: I46ee58bee48c296753aede85a44bb007a25b4ae2