[unittest] Run an exception handler thread so crashes don't bubble up to system handler.
To enable the exception handler, a test should be run
with RUN_TEST_ENABLE_CRASH_HANDLER.

Tests can then call REGISTER_CRASH to register any
processes that are expected to crash.

If a crash was registered and occurs, the exception is
caught and the test resumes.

If a crash was not registered, the exception will be passed up to
crashlogger and then the test will fail immediately.

At the conclusion of the test, if a crash was registered
and did not occur, the test will fail.

MG-802

Change-Id: Ie1a7dd3db33109e4c5ea9cfef75097df0365b35b