[audio_server] Permit AudioRenders to change MediaType.
Make audio format information an property of the links betwee audio
renderer and audio output instead of having audio outputs query the
audio renderer for information related to the format directly.

This allows audio renderers to change format without needing to hold
off a mixer who is currently in the process of mixing simply by
invalidating existing links (to be cleaned up by the outputs when they
are ready) and creating new ones with the new format.

Note: right now, we will deny a format change if there is still
pending data to be rendered queued on existing links.  Technically,
however, this would still be an OK thing to do.  The only problem
right now is that the renderer would never be able to flush the data
on the old links if it needed to (which is why we disallow the
operation).

MTWN-8 #done

Change-Id: Ibfe6acf5064969e05c32a94f656e104dbf37cf46