Record context for stories.
StoryProviderImpl holds a maxwell::ContextListener through an
instance of ContextHandler that keeps a current copy of the
context by listening to context updates.

Whenever a Story is created or focused, the current context
is recorded with the story. To accomplish this, a connection
to the focus service is passed by user runner impl to
story provider impl.

All writing to the story page is delegated to story storage
impl.

Collateral fix: Confine reading of all module data from the
story page to the Module/ prefix. Previously it was read from
a full snapshot starting at the Module/ prefix, which happened
to result in only the module data by accident.

Next up: Whenever context changes, compute an "importance"
value for each story based on the similarity of the context
the story was created in from the current context.

Change-Id: I9eeb5a7f7ede644dacc0ea39690b9aa8a2028f87