Add module_path to AddModule() and GetLink() in StoryController
Clients of StoryController can now specify a particular module path
they want to be the parent of the added module and link when they call
AddModule(). Correspondingly they can provide a module_path to GetLink()
to retrieve that created link.

This allows StoryShell to display externally added modules in a
meaningful relationship with modules already present.

Change-Id: I60d1d72f2e539bedff7a805a293cbe5bffe9aa10