[futexes] Fix restarting of futex_wait() when thread is suspended and resumed
Before this change, suspending and resuming a thread that is waiting
in the mx_futex_wait() syscall would cause mx_futex_wait() to return
ERR_TIMED_OUT (instead of restarting the syscall such that it
continues to wait).

That happened because FutexWait() was returning ERR_TIMED_OUT instead
of propagating ERR_INTERRUPTED_RETRY (which would cause the VDSO's
syscall wrapper to retry the syscall).  The code for doing that
predated the addition of thread suspension and ERR_INTERRUPTED_RETRY.

Fix this by propagating the right error code.

MG-795

Change-Id: Idd6c5c144642e9b359d54069aef4ae9c465ddb0d