Fix fcp to remove file on remote end.
fcp was using "! rm -f ${dst}" to remove the file before putting
the new one.  This is necessary because the put will fail if the
destination in on bootfs (i.e. /system).  However the "!" was
causing the rm command to run on the local side.  I'm assuming the
intention here was to not fail if the remote file does not exist.
The "-" prefix does that.  Additionally sftp's rm command does not
accept the -f option.

Change-Id: I88eda8635f78a7ca2850394401a3d2b9d70084c1