[kernel] Reduce duplication in checking THREAD_SIGNAL_KILL
Move the duplicated code for checking and handling THREAD_SIGNAL_KILL
into a function.

On the one hand, passing spin_lock_saved_state_t here is somewhat
awkward, and this is more lines of code.  On the other hand, it is
good to make it more explicit that this check happens in multiple
places.

Change-Id: Idb1c4f993729fd961ee2212c2b5ab8c6be5d13e9