[kernel][arch][x86] Revamp stack-protector/safe-stack early setup
This moves the setup of the boot CPU's x86_percpu in %gs.base
into assembly and the most basic initialization of its x86_percpu
into a compile-time static initializer.  This makes it possible
for all the C code on the startup path to use safe-stack and/or
stack-protector (which gets harder and harder to avoid with LTO).

Now the only C function that runs at all is x86_secondary_entry.
The work it does to discover its CPU number (and thus the right
address for its own x86_percpu) is too complex to do in assembly.
Instead, x86_secondary_entry is split so that the first C
function run on secondary CPUs is a simple one that manages to
avoid stack-protector/safe-stack code, and it sets up %gs.base
and enough of x86_percpu that it can then call the second C
function and that function can safely use safe-stack and/or
stack-protector code.

Change-Id: If8f0860a05a0bd98285a15d0f2f97699bf16d9d7