[kernel] Add embedded version info and kernel-buildsig tool to extract it
This embeds a simple signature early in the kernel binary that
can be searched for in raw binary memory to identify the kernel
version information.  The new tool kernel-buildsig extracts this
from a file (which can be magenta.bin or a raw memory dump, or
an ELF file, or anything else containing the right chunk).

Change-Id: I250b277f49563ee81abef16b937ff7b2a22e8dc0