[kernel][debug] make sure the scheduler gets a chance to run after a few debug commands
Both k threads list and k threadload run a set of printfs inside the
THREAD_LOCK, which though it queues up text in the debuglog does not get
to reschedule afterwards which sometimes leads to long delays before the
text shows up.

Force a reschedule after the printfs to make sure things are working.

Change-Id: I94c12ef83b7c3bb3d91c4076c4036715e08f96b9