[devmgr] Run appmgr directly
After this patch, devmgr runs appmgr directly rather than bouncing through
init. The init binary wasn't adding any value to the system. It's work can just
be done by devmgr directly and we can save resources.

As a result of this patch, the |fuchsia| job is now a peer with the
|magenta-drivers| and |magenta-services| jobs, which makes more logical sense.

Finally, this patch reduces the rights that child processes are given to the
jobs in which they run. This prevents, for example, a devhost from renaming the
|magenta-drivers| job.

Change-Id: I4db5b4338c7fcb684fbbb7074537d3cc652240d2