[hid] Raise the hardcoded report limit.
Raise the limit on the number of unique report IDs which may exist in
a HID report descriptor.  The old limit was 16, but some devices have
more.  This CL raises the limit to 32.

Note, there really should not be a limit here; eventually when we
implement a generic hid report parsing library, this limit should be
removed entirely.

Change-Id: I287fabf2874176604ed52a8c1350cc174b05e6d2