[i2c-hid] Fix an underflow.
When speculatively reading i2c-hid reports (we are currently polling
right now, apparently there is no support for GPIO sourced interrupts
yet), make sure we sanity check the report length before passing the
report on to the hid driver.  The report needs to be at least 2 bytes
long in addition to being <= the size of the buffer we read into.

Change-Id: I4621c2b331c4f1c5c1007179e349f2764f492916