Fix failed test count
The previous change to add "empty" test results wasn't incrementing the
number of failed tests.

The failed test count was being tracked by checking the length of a list
of failed test results, which are also used for a summary of failures.
Since empty results aren't easy to fit into the same format as full
results (for example, they lack elapsed time), the easiest temporary fix
is to keep separate passed and failed counts, independent of the list of
failed results. This isn't optimal for long-term usability, because the
empty results won't be included in the summary.

Also add some test coverage for Driver.print_summary().

Change-Id: Ifa0503a979d902383ed9869e8ae47cff27ce7f1b