[ulib][musl][ldso][asan] Add no_sanitize attribute on early startup path
Annotate functions on the early startup path of the dynamic
linker to disable ASan code generation.  This goes on the code
that has to run before the shadow memory is set up.  Once vDSO
symbol references are resolved, the dynamic linker can set up the
shadow memory and then the rest of the dynamic linker's own code
can be sanitized and run safely without crashing or calling any
ASan runtime functions (unless it has a bug).

Change-Id: Icbe214b5e51381c0773f71cb05d1674e4b4da1d1