Compute importance of stories.
The more similar the context in which a story is usually used to
the current context, the more important it is in the current
context.

In this implementation, no caching nor precomputation is applied
to the computation of importance of stories on the timeline. There
are design measures in place to limit resource waste:

* Clients track importance changes through a dedicated watcher. So
  no client that is current not interested in importance will get
  callbacks that accidentally trigger importance computation.

* Tracking of context (and corresponding importance change
  notifications) is confined to the dimensions of context that
  actually influence importance.

* Logging of story context, listening for context changes, and
  computation of story importance are all confined to only the
  topics thatare used to influence story importance for now.

The computation of importance is done by the story controller,
using log data read from story storage. It is only computed for
stories that have a story controller. This is ok for now, because
armadillo creates a story controller for all stories, but later
needs to be changed to compute importance for all stories (in
conjunction with caching / precomputation).

Add a test user shell that verifies that importance can be
computed, and is influenced by the context a story was created in.

Left to do in followup:
* Receive importance changed notifications.
* Importance is influenced by focus too.

Collateral fix:
* hello_world_flutter now calls ModuleContext.Ready() in
  its Module.Initialize() method.

Change-Id: Ib2c3ac4150a89e8df3c6ff78e37ff3d4a0a20379