[device][vfs] improved directory watch protocol (1 of n)
The watch protocol is revised to support more events than
just "file added" and to allow for multiple events to be
delivered in a single message.

This patch upgrades existing file watch server and client
code to use the new event format, but does not yet make use
of multiple events per message or events other than file
added.

The new IOCTL_VFS_WATCH_DIR_V2 allows for an event mask to
be specified and makes the connection process pipelined.
It is not yet implemented by servers or clients.  It will
eventually subsume the old version and drop the _V2.

mxio watcher's mxio_watcher_wait() is removed as it is
unused and its API is not suitable for the eventual multiple
events per message model.

Change-Id: Ia9fa3a4649ae887c6de1894784e7d8f3dbf880db