[fs][minfs][blobstore] Don't #define info
Previously, trace.h defined info, warn, and error as macros. These names are
very commonly used and as this header is being included in more places these
macros caused name conflicts.

This patch changes the macros to be in ALL_CAPS and to start with "FS_" to try
to avoid conflicts with other macros.

Change-Id: Idd9e27efac57624db530b0b1b1e81a934da22f3b