[ddktl] Relax a ddktl assert.
If a ddk::Device<> inherits from base_protocol, demand that it has a
valid protocol ID at device add time, but do not demand that it have a
protocol thunk table.  It is legal to have a protocol (so a driver
ends up in the proper /dev/class location) but no in-process thunks.

Change-Id: I376171c74aa442a90596c6239f7ac9249de456f0