[guest] Boot linux from bzImage, setup zeropage.
This change brings console output to linux, as well as
further boot progress, which is pretty satisfying.

Stop using uncompressed vmlinux and use the bzImage instead.
The bzImage already has a partially correct zeropage for the
kernel AND a decompressor, so we will get sensible defaults,
decompression and a single, small image file.

Tweak UART emulation for linux compatibility.

Change-Id: I29ed369a2827731ba5d4d7666b7d731289ae8b9a