[kernel][magenta] remove 'strobing' test for timers
Periodic timers strobe the signal, which reliable wakes up
the waiters but it does not have a reliable 'moment in time'
when the signal is true.

More specifically, it is true when the waiter is awake and
should appear in |pending| but after that reading it is racy.

Waiting on it again is not racy though.

Change-Id: I6530651f1be8684f08b8d6e25e0bc741035a56d1