Introduce Timer objects
A Timer wraps an mx::timer object and allows components to manage their
own timeouts. The Device (via DeviceInterface) produces Timers after
registering them with the port. The MLME will generate ids and track the
pairing of the id and the owner of the Timer, and the ids are used to
notify components when the Timer has fired.

A TestTimer may be used in unittests; it does nothing on Start/Cancel,
but does reference a TestClock which can be set to any absolute time.

Change-Id: I2b79dca606914c8c3993e1f0629e1de6196bb5ff