[merkle] Fix length calculation for small data.
This CL fixes a bug introduced by 65e16c9.  A nonzero data-length led
VerifyRoot to assume the root of the Merkle tree was eactly one node
long.  This true if there is any tree at all, but if the data fits in a
single node, we shouldn't pad the length.

An updated unit test for this condition is being omitted for now to
expedite this fix.  It will be in a following CL.

Change-Id: I49d8e8344bd3e9159a677af0e05f91173f289805