[merkle] Refactor unit tests.
This CL adds the unit test alluded to in CL 33688. It adds unit tests
and test values for half a node of data.

It furthermore expands existing unit tests to handle additional lengths
by splitting then into helper functions and a loop that iterates of the
following data lengths: 0, 1, kNodeSize/2, kNodeSize-1, kNodeSize,
kNodeSize+1, kNodeSize*8, kNodeSize*257, and kNodeSize*515/2.

There is some incidental formatting and cleanup.

Finally, it fixes a one-line bug in VerifyLevels in which 'length' was
being computed incorrectly.  This bug was exposed by the added tests.

Change-Id: Ic0b45119fd7edd5ae3729022011b3119e0291101