Fix the FTL_DCHECK logic in auth.
In release builds, credentials file was not getting written on successful login
attempts because of the omission of this FTL_DCHECK.

FTL_DCHECK happens only on debug builds so's it is compiled out on release
builds. As a result the outcome of writing the response after successful
login is omitted on release builds although it was a successful
operation. This cl does the FTL_DCHECK on a local variable after getting
back the response first.

Change-Id: Ifffd41fb5702f2d1e3de019e8cfc57a21931f616