Add symbol_visibility_hidden to default for all targets
The previous logic added -fvisibility=hidden to cflags only for
shared_library and loadable_module targets, but not source_set or
static_library targets. This resulted in very confusing behavior when
moving code between shared_library() and source_set()s. Since this flag
controls the behavior when compiling files, not linking the object
files together, having it specified only on the link targets doesn't
make much sense.

This makes the default hidden for all targets by default. To change the
default for a particular target, add

  configs -= [ "//build/config:symbol_visibility_hidden" ]

to the appropriate target in a BUILD.gn file.

Change-Id: I078c16cf19b5088e9a33727c4ffc4bf2e3bbbf09