[mxtl] Never let a class size depend on DEBUG v. RELEASE
Eliminate instances of classes in mxtl which have data members which
may be present or not based on the current preprocessor value of
MX_DEBUG_ASSERT_IMPLEMENTED.  #defs are brittle, and when two pieces
of code running in the same process share a mxtl object, but each
piece has been compiled with different values for the #def, it can
lead to code accessing the wrong parts of an object if the code
accessing the object disagrees with the code which allocated the
object.

Eliminate the instaces of this pattern in mxtl by doing one of two
things.

1) Where possible, just eliminate the need for the conditional state.
2) Where this is not possible, make the presence or absence of the
   state a property of the type of the object.  If there is a
   disagreement about the nature of an object (debug v. release) and
   therefor its size, it should be caught at compile time due to a
   type mismatch instead of caught at runtime due to strange access
   patterns.

MG-838 #done

Change-Id: Idfad5214444a471bdd28905f48848e68a239459a