[musl][pthread][ldso] Improve default thread names and stack VMO names
When a thread has an explicit name, give its TCB/stack VMO a name
derived from that name but that also includes the thrd_t (aka
pthread_t) value if it fits.  Also include the TLS size if that
fits, since it's the piece of that VMO's size calculation that
you can't see explicitly from the mapping sizes (the other two
mappings are exactly the stack size, while the TCB mapping's size
is rounded up from the TLS size plus internal state).

For threads with no explicit name, give both the VMO and the
thread a default name that includes the same information.

Change-Id: Id3b5986a56c3322c8f073a6fff31ca7ecc658255