Refactor test results reporting
I found this area of the code hard to follow, because every method
was intended to run on a specific thread, and this would sometimes be
different threads for the same object.

Also, a single class was responsible for both listening to events from
the GoogleTest framework and reporting them to the TestRunner service.

I've factored things out into three classes:
- GTestListener handles events from GoogleTest: individual test results
  and the completion of the test program. It builds a message string for
  each test failure and creates a TestResult object.
- ResultsQueue is a thread-safe queue for passing TestResult objects
  from GTestListener to Reporter. It uses a Magenta event object for
  signaling, in order to be compatible with mtl::MessageLoop.
- Reporter runs an mtl::MessageLoop (because ApplicationContext
  complains if there isn't one, and because the Tracing framework
  uses it), and it registers itself as an MessageLoopHandler for when
  there are TestResult object on the queue. When it gets null, it calls
  Teardown on TestRunner, waits for the acknowledgement callback, and
  quits the message loop.

Change-Id: I3c979b25221b8bf14e77e001e6e32bd8b0799ac0