When re-inflating stories use module_paths from story storage.
The earlier implementation was passing an empty module_path for all
modules started directly by framework. After some recent changes it is
possible that TQI creates a non-root module who's lifecycle is to be
managed by the framework but is not root in the surface sense of things.

Follups:
(1) Register a module watcher for all modules who's lifecycle is managed
by the framework and update story state appropriately.
(2) TBD StoryController.AddModule() should not be able to add a child to
existing module (child in a framework sense of things via module_path)
but rather indicate a child (child in a surface sense of things) which
is kept track of in module data.

Change-Id: I12faf3b96277582cc705904a73b08f191c3da8a9