[udev][usb-mass-storage] Wait for CBW to complete before transferring data
The CBW is the USB mass storage command block that is sent at the beginning of
all UMS transactions. When doing bulk data writes, we have been sending the
data payload immediately after the CBW. On some USB drives, we are seeing
sporadic failures where the data transfer fails with "not ready".
Waiting for the CBW transfer to be ACKed before starting the data transfer
seems to fix this problem.

We probably have more work to do here, but this seems to fix our immediate issue.

US-263 #done

Change-Id: If665e1e4914891228e42a165545f0e4bd701d59e