Remove MessageQueue.Receive(), use a MessageReader interface instead.
The MessageQueue.Receive() mechanism had the acknowledgement problem:
The MessageQueue interface could die with a message pending in the pipe
(during a Module or Agent being stopping), and that message will never
be re-sent again. To fix this, this CL inverts the mechanism:

We now use an interface that requires a user to 'ack' the message, or
otherwise it will be resent when another watcher is registered. The
restriction being that there can only be one registered receiver. There
can still be multiple senders.

Note that the actual fidl interface is called MessageReader, not
MessageReceiver, since 'MessageReceiver' is a conflicting class name in
the fidl C++ bindings; once this is fixed, we can rename this fidl
interface to MessageReceiver.

This CL (and accompanying CLs) also comes with a
'MessageReceiverAdapator' class in C++ and Dart to make dealing with
this interface easier.

Change-Id: I78a6b9bc28a592650741df1c490e8927390c13d7