Preload StoryShell Application
Since all stories in a user_runner use the same StoryShell
implementation, we preload the StoryShell application for faster story
start times. There is a memory cost of one extra StoryShell process
running, but this is a worthwhile tradeoff in the short term.

In order to do this we need to proxy the ViewOwner for an indefinite
amount of time.

Change-Id: I9bd5865a2348295050d46f518454ea632add5978