Classify modules as INTERNAL / EXTERNAL and use it to drive state.
NOTE: This is an intermediate solution for demo.

A module is said to be INTERNAL if it was started from within the story
via ModuleContext.StartModule* .
A module is said to be EXTERNAL if it was started from outside the story
via StoryController.AddModule or if it's the first module in the story.

We save this information along with ModuleData.

When starting for the first time or re-inflating, the module controllers
for all EXTERNAL modules are owned by the framework which registers a
watcher. On receiving a DONE state, the module is de-focussed in story
shell and stopped (not actually stopped due to a bug in story shell).

Currently, the StoryState is still governed by the state of the root
module.

Change-Id: Ib488639eca5374b3a9b144d200eb3e9b2bac655b